Move FilterSlots, NbtCompat, DirectionSerializer out of core.lib.storage
