+++ common/src/main/java/com/logistics/automation/marker/MarkerBlockEntity.java
@@ -2,7 +2,7 @@ package com.logistics.automation.marker;
 
 import com.logistics.LogisticsAutomation;
 import com.logistics.core.lib.BaseBlockEntity;
-import com.logistics.core.lib.storage.NbtCompat;
+import com.logistics.core.lib.compat.NbtCompat;
 import java.util.ArrayList;
 import java.util.List;
 import net.minecraft.core.BlockPos;
