+++ common/src/main/java/com/logistics/core/lib/items/ItemInventoryComponent.java
@@ -1,6 +1,6 @@
 package com.logistics.core.lib.items;
 
-import com.logistics.core.lib.storage.NbtCompat;
+import com.logistics.core.lib.compat.NbtCompat;
 import net.fabricmc.fabric.api.transfer.v1.item.ContainerStorage;
 import net.fabricmc.fabric.api.transfer.v1.item.ItemVariant;
 import net.fabricmc.fabric.api.transfer.v1.storage.Storage;
