+++ common/src/main/java/com/logistics/pipe/modules/AdvancedExtractorModule.java
@@ -6,8 +6,8 @@ import com.logistics.pipe.network.NetDbg;
 import com.logistics.core.lib.pipe.Module;
 import com.logistics.core.lib.pipe.TickingModule;
 import com.logistics.core.lib.resource.ResourceId;
-import com.logistics.core.lib.storage.DirectionSerializer;
-import com.logistics.core.lib.storage.FilterSlots;
+import com.logistics.core.lib.serialization.DirectionSerializer;
+import com.logistics.core.lib.filter.FilterSlots;
 import com.logistics.core.lib.pipe.PipeContext;
 import com.logistics.pipe.block.entity.PipeBlockEntity;
 import com.logistics.core.lib.pipe.TravelingItem;
