+++ common/src/main/java/com/logistics/pipe/modules/EnchantmentSinkModule.java
@@ -7,7 +7,7 @@ import com.logistics.pipe.network.NetDbg;
 import com.logistics.core.lib.pipe.ItemAcceptingModule;
 import com.logistics.core.lib.pipe.Module;
 import com.logistics.core.lib.resource.ResourceId;
-import com.logistics.core.lib.storage.DirectionSerializer;
+import com.logistics.core.lib.serialization.DirectionSerializer;
 import com.logistics.core.lib.pipe.PipeContext;
 import com.logistics.core.lib.network.ILogisticsNetwork;
 import com.logistics.core.lib.pipe.RoutePlan;
