+++ common/src/main/java/com/logistics/pipe/modules/ModSinkModule.java
@@ -6,7 +6,7 @@ import com.logistics.LogisticsPipe;
 import com.logistics.core.lib.pipe.*;
 import com.logistics.core.lib.pipe.Module;
 import com.logistics.core.lib.resource.ResourceId;
-import com.logistics.core.lib.storage.DirectionSerializer;
+import com.logistics.core.lib.serialization.DirectionSerializer;
 import com.logistics.pipe.block.entity.PipeBlockEntity;
 import com.logistics.core.lib.network.ILogisticsNetwork;
 import com.logistics.pipe.network.NetDbg;
