+++ common/src/main/java/com/logistics/pipe/modules/ProviderModule.java
@@ -9,8 +9,8 @@ import com.logistics.core.lib.pipe.Module;
 import com.logistics.core.lib.pipe.TickingModule;
 import com.logistics.pipe.network.NetDbg;
 import com.logistics.core.lib.resource.ResourceId;
-import com.logistics.core.lib.storage.FilterSlots;
-import com.logistics.core.lib.storage.NbtCompat;
+import com.logistics.core.lib.filter.FilterSlots;
+import com.logistics.core.lib.compat.NbtCompat;
 import com.logistics.core.lib.pipe.PipeContext;
 import com.logistics.pipe.network.NetworkRegistry;
 import com.logistics.core.lib.pipe.TravelingItem;
