+++ common/src/main/java/com/logistics/pipe/modules/RequesterModule.java
@@ -7,8 +7,8 @@ import com.logistics.core.lib.network.ILogisticsNetwork;
 import com.logistics.core.lib.pipe.Module;
 import com.logistics.core.lib.pipe.TickingModule;
 import com.logistics.core.lib.resource.ResourceId;
-import com.logistics.core.lib.storage.DirectionSerializer;
-import com.logistics.core.lib.storage.NbtCompat;
+import com.logistics.core.lib.serialization.DirectionSerializer;
+import com.logistics.core.lib.compat.NbtCompat;
 import com.logistics.core.lib.pipe.PipeContext;
 import com.logistics.pipe.block.entity.PipeBlockEntity;
 import com.logistics.pipe.network.NetworkRegistry;
