+++ common/src/main/java/com/logistics/pipe/modules/SinkModule.java
@@ -8,8 +8,8 @@ import com.logistics.core.lib.pipe.Module;
 import com.logistics.pipe.network.NetDbg;
 import com.logistics.core.lib.network.ILogisticsNetwork;
 import com.logistics.core.lib.resource.ResourceId;
-import com.logistics.core.lib.storage.DirectionSerializer;
-import com.logistics.core.lib.storage.FilterSlots;
+import com.logistics.core.lib.serialization.DirectionSerializer;
+import com.logistics.core.lib.filter.FilterSlots;
 import com.logistics.pipe.block.entity.PipeBlockEntity;
 import com.logistics.core.lib.pipe.RoutePlan;
 import com.logistics.core.lib.pipe.TravelingItem;
