+++ common/src/main/java/com/logistics/pipe/ui/SupplyInventory.java
@@ -2,7 +2,7 @@ package com.logistics.pipe.ui;
 
 import com.logistics.core.lib.resource.ResourceId;
 import com.logistics.core.lib.pipe.PipeContext;
-import com.logistics.core.lib.storage.NbtCompat;
+import com.logistics.core.lib.compat.NbtCompat;
 import com.logistics.pipe.block.entity.PipeBlockEntity;
 import com.logistics.pipe.modules.SupplierModule;
 import net.minecraft.core.NonNullList;
