+++ common/src/main/java/com/logistics/power/block/entity/CreativeSinkBlockEntity.java
@@ -4,7 +4,7 @@ import com.logistics.LogisticsPower;
 import com.logistics.core.lib.BaseBlockEntity;
 import com.logistics.core.lib.block.capability.HasEnergyStorage;
 import com.logistics.core.lib.power.AcceptsLowTierEnergy;
-import com.logistics.core.lib.storage.NbtCompat;
+import com.logistics.core.lib.compat.NbtCompat;
 import com.logistics.core.lib.support.ProbeResult;
 import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
